+++ config/test.exs
@@ -1,13 +1,15 @@
 use Mix.Config
 
 config :k8s,
-  discovery: %{
-    driver: K8s.Discovery.Driver.File,
-    opts: [config: "test/support/discovery/example.json"]
-  },
+  discovery_driver: K8s.Discovery.Driver.File,
+  discovery_opts: [config: "test/support/discovery/example.json"],
   http_provider: K8s.Client.DynamicHTTPProvider,
   clusters: %{
     test: %{
-      conn: "test/support/kube-config.yaml"
+      conn: "test/support/kube-config.yaml",
+      conn_opts: [
+        discovery_driver: K8s.Discovery.Driver.File,
+        discovery_opts: [config: "test/support/discovery/example.json"]
+      ]
     }
   }