+++ lib/k8s/discovery/driver/file.ex
@@ -71,8 +71,8 @@ defmodule K8s.Discovery.Driver.File do
 
   @spec default_opts() :: Keyword.t()
   defp default_opts() do
-    case K8s.default_driver() do
-      __MODULE__ -> K8s.default_driver_opts()
+    case K8s.Discovery.default_driver() do
+      __MODULE__ -> K8s.Discovery.default_opts()
       _ -> []
     end
   end